New vision of the project

+++ backend/src/service/getfile.go
@@ -3,7 +3,7 @@ package service
 import "models"
 
 func GetFile(id int) *models.Upload {
-	return models.GetFile(id)
+	return models.GetFile(int64(id))
 }
 
 func GetFiles() []models.Upload {
